// Front/ProEventos-App/src/app/models/Evento.ts
import { Lote } from "./Lote";
import { PalestranteEvento } from "./PalestranteEvento";
import { RedeSocial } from "./RedeSocial";

export interface Evento {
 id: number;
 local: string;
 dataEvento?: Date;
 tema: string;
 qtdPessoas: number;
 imagemURL: string;
 telefone: string;
 email: string;
 lotes: Lote[];
 redeSociais: RedeSocial[];
 palestranteEventos: PalestranteEvento[];
}